requirements.txt
mcp
PyYAML
lxml
paramiko
uvicorn
starlette
